Moved handle out of encode and into its own section
