+++ e2e/e2e_test.go
@@ -16,6 +16,8 @@ import (
 	"github.com/cwxstat/rabbitmq/lib/conn"
 	"github.com/cwxstat/rabbitmq/lib/flag"
 
+
+	"github.com/cwxstat/rabbitmq/lib/handle"
 	"github.com/cwxstat/rabbitmq/lib/consumer"
 	"github.com/cwxstat/rabbitmq/lib/producer"
 	//"github.com/cwxstat/rabbitmq/e2e"
@@ -101,7 +103,7 @@ var _ = Describe("E2e", func() {
 				Expect(err).To(BeNil())
 				os.RemoveAll("/tmp/compress.tar.gz")
 
-				handler := &encode.HS{}
+				handler := &handle.HS{}
 				handler.File = "/tmp/junk.tar.gz"
 				handler.DestDir = desDir
